Use fixed width integers by default and update tests

+++ Tests/Support/Snapshots/edgecases-data-types/Sources/Entities/FormatTest.swift
@@ -9,7 +9,7 @@ public struct FormatTest: Codable {
     public var int32: Double?
     public var int64: Int?
     public var number: Double
-    public var float: Double?
+    public var float: Float?
     public var double: Double?
     public var string: String?
     public var byte: String
@@ -19,7 +19,7 @@ public struct FormatTest: Codable {
     public var uuid: UUID?
     public var password: String
 
-    public init(integer: Int? = nil, int32: Double? = nil, int64: Int? = nil, number: Double, float: Double? = nil, double: Double? = nil, string: String? = nil, byte: String, binary: String? = nil, date: NaiveDate, dateTime: AnyJSON? = nil, uuid: UUID? = nil, password: String) {
+    public init(integer: Int? = nil, int32: Double? = nil, int64: Int? = nil, number: Double, float: Float? = nil, double: Double? = nil, string: String? = nil, byte: String, binary: String? = nil, date: NaiveDate, dateTime: AnyJSON? = nil, uuid: UUID? = nil, password: String) {
         self.integer = integer
         self.int32 = int32
         self.int64 = int64
@@ -41,7 +41,7 @@ public struct FormatTest: Codable {
         self.int32 = try values.decodeIfPresent(Double.self, forKey: "int32")
         self.int64 = try values.decodeIfPresent(Int.self, forKey: "int64")
         self.number = try values.decode(Double.self, forKey: "number")
-        self.float = try values.decodeIfPresent(Double.self, forKey: "float")
+        self.float = try values.decodeIfPresent(Float.self, forKey: "float")
         self.double = try values.decodeIfPresent(Double.self, forKey: "double")
         self.string = try values.decodeIfPresent(String.self, forKey: "string")
         self.byte = try values.decode(String.self, forKey: "byte")
